### 2024-05-17 327ecd4
more ctor checks in AeroMode, sampled test for AeroDist
