### 2024-05-17 18b3c2f
addressing pylint hints
