### 2024-05-17 7069e0c
skip exception tests on Apple Silicon
