### 2024-05-21 201d2e6
handle multiple calls to read_named_array to fix sampled mode reading
