### 2024-05-21 5a45044
simpler cast
