### 2024-05-22 4aecd30
add getters for sampled aero_mode properties. add to tests
