### 2024-05-22 ab6b60b
add dist sample tests for sampled and monodisperse modes
