### 2024-05-22 2f155e9
remove code duplication in tests (thank you pylint)
