### 2024-05-22 d6dbe31
support for AeroMode instantiation with sampled mode type (incl. fixes in spec file logic) + sampled and mono mode types coverage in unit tests (in test_aero_[mode,state,dist].py) (#357)
Co-authored-by: Jeffrey Curtis <jcurtis2@illinois.edu>
