### 2024-05-25 #179 @Batalex
Hey @slayoo, I'm Alex, currently serving as the Editor-in-Chief. I'm sorry it took me so long to get back to you.

`PyPartMC` seems to be in scope for us. I truly appreciate the care you put into the submission, especially the comparative section. It's very valuable to a profane such as myself when it comes to evaluating submissions.

Please find below the preliminary checks.

## Editor-in-Chief checks

Thank you for submitting your package for pyOpenSci review. Below are the basic checks that your package needs to pass to begin our review. If some of these are missing, we will ask you to work on them before the review process begins.

Please check our [Python packaging guide](https://www.pyopensci.org/python-package-guide) for more information on the elements below.

  **The package installation does not install the dependencies**

---
---

I have a few suggestions regarding the documentation:
- Having the links to the binder/collab/etc. is great. Would you consider including those end to end examples in the [documentation website](https://open-atmos.github.io/PyPartMC/)? It is a pity that it only has the API reference. I'd like to see more textual content as well, to properly frame the code blocks and fully understand what is going on
- I think it would be great to have a dedicated CONTRIBUTING.md/rst file in the repository with the README's sections about the implementation details and the developers notes.
- We suggest adding a code of conduct document as well. It's always a good idea to set the proper expectations from the start!
