### 2024-05-26 #179 @slayoo
@Batalex, thanks for the feedback!
We've just started addressing the above points with a code-of-conduct PR: https://github.com/open-atmos/PyPartMC/pull/358
I'll report back here after addressing all points
