### 2024-05-28 202a02b
Create CODE_OF_CONDUCT.md (#358)
