### 2024-05-29 f3479db
improve pdoc landing site by rendering the README contents in the index.html. closes #332
