### 2024-05-29 79a4b62
move developer notes into a new CONTRIBUTING.md file
