### 2024-05-29 10120b7
move developer notes into a new CONTRIBUTING.md file (#360)
