### 2024-05-29 #179 @slayoo
@Batalex, with the following PRs just merged:
- https://github.com/open-atmos/PyPartMC/pull/358
- https://github.com/open-atmos/PyPartMC/pull/359
- https://github.com/open-atmos/PyPartMC/pull/360

the repo and docs are improved following your suggestions:
- https://open-atmos.github.io/PyPartMC
- https://github.com/open-atmos/PyPartMC/tree/main/CODE_OF_CONDUCT.md
- https://github.com/open-atmos/PyPartMC/tree/main/CONTRIBUTING.md
