### 2024-05-31 #179 @Batalex
Thank you, @slayoo, the docs look great!
`PyPartMC` is a strong submission, I'll get started on finding an editor right now 🫡
