### 2024-05-31 #179 @Batalex
Sorry, I forgot to add something important to my previous message. Since our field of expertise is Python, the significant portion of C++ code in `PyPartMC` will only be evaluated on the best effort basis (which means that maybe it won't be evaluated). Our focus will be on the Python packaging side of the project.
