### 2024-06-04 f8496e6
add consistency checks for allow halving/doubling flags
