### 2024-06-05 c9b40e4
disable new exception-checking tests on Apple Silicon
