### 2024-06-06 777c8bd
first pass at linking camp
