### 2024-06-06 c31dbd1
first pass at linking camp
