### 2024-06-15 8716be5
add (mostly) empty constructors for data types
