+++ tests/test_env_state.py
@@ -34,6 +34,14 @@ class TestEnvState:
         assert sut is not None
 
     @staticmethod
+    def test_ctor_empty():
+        # act
+        sut = ppmc.EnvState()
+
+        # assert
+        assert sut is not None
+
+    @staticmethod
     def test_dtor():
         # arrange
         sut = ppmc.EnvState(  # pylint: disable=unused-variable
