### 2024-06-18 186eced
add consistency checks for allow halving/doubling flags. closes #266 (#362)
