+++ tests/test_run_part.py
@@ -4,6 +4,8 @@
 # Authors: https://github.com/open-atmos/PyPartMC/graphs/contributors                              #
 ####################################################################################################
 
+import platform
+
 import numpy as np
 import pytest
 
@@ -52,7 +54,7 @@ class TestRunPart:
 
     @staticmethod
     def test_run_part_timestep(common_args):
-        (last_output_time, last_progress_time, i_output) = ppmc.run_part_timestep(
+        last_output_time, last_progress_time, i_output = ppmc.run_part_timestep(
             *common_args, 1, 0, 0, 0, 1
         )
 
@@ -63,14 +65,18 @@ class TestRunPart:
 
     @staticmethod
     def test_run_part_timeblock(common_args):
+        # arrange
         num_times = int(
             RUN_PART_OPT_CTOR_ARG_SIMULATION["t_output"]
             / RUN_PART_OPT_CTOR_ARG_SIMULATION["del_t"]
         )
-        (last_output_time, last_progress_time, i_output) = ppmc.run_part_timeblock(
+
+        # act
+        last_output_time, last_progress_time, i_output = ppmc.run_part_timeblock(
             *common_args, 1, num_times, 0, 0, 0, 1
         )
 
+        # assert
         assert last_output_time == RUN_PART_OPT_CTOR_ARG_SIMULATION["t_output"]
         assert last_progress_time == 0.0
         assert i_output == 2
@@ -94,8 +100,65 @@ class TestRunPart:
                 "do_condensation": True,
             }
         )
-        aero_state.dist_sample(aero_dist, 1.0, 0.0, True, True)
+        aero_state.dist_sample(aero_dist, 1.0, 0.0, False, False)
         ppmc.condense_equilib_particles(env_state, aero_data, aero_state)
         ppmc.run_part(*args)
 
         assert np.sum(aero_state.masses(include=["H2O"])) > 0.0
+
+    @staticmethod
+    @pytest.mark.parametrize(
+        "flags",
+        (
+            ((True, True), (True, False)),
+            ((True, True), (False, True)),
+            ((True, True), (False, False)),
+            ((False, False), (True, False)),
+            ((False, False), (False, True)),
+            ((False, False), (True, True)),
+            ((True, False), (False, False)),
+            ((True, False), (False, True)),
+            ((False, True), (False, False)),
+            ((False, True), (True, False)),
+        ),
+    )
+    @pytest.mark.parametrize(
+        "fun_args",
+        (
+            ("run_part", []),
+            ("run_part_timestep", [0, 0, 0, 0, 0]),
+            ("run_part_timeblock", [0, 0, 0, 0, 0, 0]),
+        ),
+    )
+    @pytest.mark.skipif(platform.machine() == "arm64", reason="TODO #348")
+    def test_run_part_allow_flag_mismatch(common_args, tmp_path, fun_args, flags):
+        # arrange
+        filename = tmp_path / "test"
+        env_state = ppmc.EnvState(ENV_STATE_CTOR_ARG_HIGH_RH)
+        aero_data = ppmc.AeroData(AERO_DATA_CTOR_ARG_FULL)
+        aero_dist = ppmc.AeroDist(aero_data, AERO_DIST_CTOR_ARG_FULL)
+        aero_state = ppmc.AeroState(aero_data, *AERO_STATE_CTOR_ARG_MINIMAL)
+        args = list(common_args)
+        args[0].init_env_state(env_state, 0.0)
+        args[1] = env_state
+        args[2] = aero_data
+        args[3] = aero_state
+        args[6] = ppmc.RunPartOpt(
+            {
+                **RUN_PART_OPT_CTOR_ARG_SIMULATION,
+                "output_prefix": str(filename),
[…]
