### 2024-06-18 e306784
first pass at returning objects
