+++ src/pypartmc.cpp
@@ -82,8 +82,8 @@ PYBIND11_MODULE(_PyPartMC, m) {
              same, but without the \c _a suffix.
         )pbdoc"
     )
-        .def(py::init<const nlohmann::json&>(), py::arg("arg")=py::dict())
-//        .def(py::init<>(), "Empty constructor for reading from file.")
+        .def(py::init<const nlohmann::json&>())
+        .def(py::init<>(), "Empty constructor for reading from file.")
         .def("spec_by_name", AeroData::spec_by_name,
              "Returns the number of the species in AeroData with the given name")
         .def("__len__", AeroData::__len__, "Number of aerosol species")
@@ -287,8 +287,8 @@ PYBIND11_MODULE(_PyPartMC, m) {
             is gas_state%%mix_rat(i).
         )pbdoc"
     )
-        .def(py::init<const py::tuple&>(), py::arg("arg")=py::tuple())
-//        .def(py::init<>(), "Empty constructor for reading from file.")
+        .def(py::init<const py::tuple&>())
+        .def(py::init<>(), "Empty constructor for reading from file.")
         .def("__len__", GasData::__len__,
             "returns number of gas species")
         .def_property_readonly("n_spec", GasData::__len__)
