+++ src/output.cpp
@@ -35,7 +35,7 @@ void output_state(
        &i_repeat, &record_removals, &record_optical);
 }
 
-std::tuple<AeroData*, GasData*> input_state(
+std::tuple<AeroData*, GasData*, EnvState*> input_state(
     const std::string &name,
     const AeroData &aero_data,
     const AeroState &aero_state,
@@ -59,9 +59,10 @@ std::tuple<AeroData*, GasData*> input_state(
     AeroState *aero_state_new = new AeroState(std::shared_ptr<AeroData>(aero_data_new));
     GasData *gas_data_new = new GasData();
     GasState *gas_state_new = new GasState(std::shared_ptr<GasData>(gas_data_new));
+    EnvState *env_state_new = new EnvState();
     f_input_state(name.c_str(), &name_size, &index, &time, &del_t, &i_repeat,
        aero_data_new->ptr.f_arg(), aero_state_new->ptr.f_arg(),
-       gas_data_new->ptr.f_arg(), gas_state_new->ptr.f_arg(), env_state.ptr.f_arg());
+       gas_data_new->ptr.f_arg(), gas_state_new->ptr.f_arg(), env_state_new->ptr.f_arg());
 
    int n_source;
    int len_gas_data;
@@ -74,5 +75,5 @@ std::tuple<AeroData*, GasData*> input_state(
    f_gas_state_len(gas_state_new->ptr.f_arg(), &len_gas_state);
    f_aero_state_len(aero_state_new->ptr.f_arg(), &len_aero_state);
 
-   return std::make_tuple(aero_data_new, gas_data_new);
+   return std::make_tuple(aero_data_new, gas_data_new, env_state_new);
 }
