### 2024-06-18 c1b75fd
clean up arguments to input_state
