### 2024-06-24 ffbf800
return aero_state and gas_state
