### 2024-06-24 6804a2c
remove empty constructors
