### 2024-06-25 e20c46f
handle aerodata and gasdata
