+++ src/output.cpp
@@ -35,7 +35,8 @@ void output_state(
        &i_repeat, &record_removals, &record_optical);
 }
 
-std::tuple<AeroData*, AeroState*, GasData*, GasState*, EnvState*, int, int> input_state(
+std::tuple<std::shared_ptr<AeroData>, AeroState*, std::shared_ptr<GasData>,
+     GasState*, EnvState*> input_state(
     const std::string &name
 ){
     int index;
@@ -44,26 +45,14 @@ std::tuple<AeroData*, AeroState*, GasData*, GasState*, EnvState*, int, int> inpu
     int i_repeat;
     const int name_size = name.size();
 
-    AeroData *aero_data = new AeroData();
     AeroState *aero_state = new AeroState(std::shared_ptr<AeroData>(new AeroData()));
-    GasData *gas_data = new GasData();
     GasState *gas_state = new GasState(std::shared_ptr<GasData>(new GasData()));
     EnvState *env_state = new EnvState();
     f_input_state(name.c_str(), &name_size, &index, &time, &del_t, &i_repeat,
-       aero_state->aero_data->ptr.f_arg(), aero_state->ptr.f_arg(),
-//       aero_data->ptr.f_arg(), aero_state->ptr.f_arg(),
-//       gas_data->ptr.f_arg(), gas_state->ptr.f_arg(),
-       gas_state->gas_data->ptr.f_arg(), gas_state->ptr.f_arg(),
-       env_state->ptr.f_arg());
+       aero_state->aero_data->ptr.f_arg_non_const(), aero_state->ptr.f_arg_non_const(),
+       gas_state->gas_data->ptr.f_arg_non_const(), gas_state->ptr.f_arg_non_const(),
+       env_state->ptr.f_arg_non_const());
 
-
-   int n_source;
-   int len_gas_data;
-
-   // Quick tests to see if everything works
-   f_aero_data_n_source(aero_state->aero_data->ptr.f_arg(), &n_source);
-   f_gas_data_len(gas_state->gas_data->ptr.f_arg(), &len_gas_data);
-
-   return std::make_tuple(aero_data, aero_state, gas_data, gas_state, env_state,
-          n_source, len_gas_data);
+    return std::make_tuple(aero_state->aero_data, aero_state, gas_state->gas_data,
+       gas_state, env_state);
 }
