### 2024-06-26 228245e
avoid Scipy 1.14.0 on CI
