### 2024-06-26 b118949
add SciPy version constraint to .binder/requirements.txt
