### 2024-06-26 9856a59
Improve input_state by returning populated data structures (#364)
Co-authored-by: Sylwester Arabas <sylwester.arabas@agh.edu.pl>
