+++ src/output.cpp
@@ -35,13 +35,9 @@ void output_state(
        &i_repeat, &record_removals, &record_optical);
 }
 
-void input_state(
-    const std::string &name,
-    const AeroData &aero_data,
-    const AeroState &aero_state,
-    const GasData &gas_data,
-    const GasState &gas_state,
-    const EnvState &env_state
+std::tuple<std::shared_ptr<AeroData>, AeroState*, std::shared_ptr<GasData>,
+     GasState*, EnvState*> input_state(
+    const std::string &name
 ){
     int index;
     double time;
@@ -49,7 +45,14 @@ void input_state(
     int i_repeat;
     const int name_size = name.size();
 
+    AeroState *aero_state = new AeroState(std::shared_ptr<AeroData>(new AeroData()));
+    GasState *gas_state = new GasState(std::shared_ptr<GasData>(new GasData()));
+    EnvState *env_state = new EnvState();
     f_input_state(name.c_str(), &name_size, &index, &time, &del_t, &i_repeat,
-       aero_data.ptr.f_arg(), aero_state.ptr.f_arg(),
-       gas_state.gas_data->ptr.f_arg(), gas_state.ptr.f_arg(), env_state.ptr.f_arg());
+       aero_state->aero_data->ptr.f_arg_non_const(), aero_state->ptr.f_arg_non_const(),
+       gas_state->gas_data->ptr.f_arg_non_const(), gas_state->ptr.f_arg_non_const(),
+       env_state->ptr.f_arg_non_const());
+
+    return std::make_tuple(aero_state->aero_data, aero_state, gas_state->gas_data,
+       gas_state, env_state);
 }
