### 2024-07-19 3781a4e
Update CITATION.cff (arXiv -> SoftwareX)
