### 2024-07-30 a8d8be0
add a basic test case draft checking parsing of chamber loss_function params
