### 2024-08-07 308ce46
add constant loss function test. give values some values to chamber
