+++ tests/test_scenario.py
@@ -9,7 +9,6 @@ import gc
 import json
 import platform
 
-import numpy as np
 import pytest
 
 import PyPartMC as ppmc
