### 2024-08-08 d00cbd8
add a basic test case draft checking parsing of chamber loss_function params (#369)
Co-authored-by: Jeffrey Curtis <jcurtis2@illinois.edu>
