### 2024-08-25 b79a584
switch from pdoc3 to pdoc, cleanup CI workflow
