### 2024-08-26 a582055
add missing camp_core
