### 2024-08-26 9a69cab
use bash shell on all platforms
