+++ .github/workflows/pdoc.yml
@@ -1,5 +1,9 @@
 name: pdoc
 
+defaults:
+  run:
+    shell: bash
+
 on:
   push:
     branches: [ main ]
