+++ .github/workflows/pdoc.yml
@@ -14,7 +14,7 @@ jobs:
   deploy:
     strategy:
       matrix:
-        platform: [ubuntu-latest, windows-latest]
+        platform: [ubuntu-latest]
     runs-on: ${{ matrix.platform }}
     steps:
       - uses: actions/checkout@v2
