### 2024-08-26 79643db
switch from pdoc3 to pdoc, cleanup CI workflow (#371)
