### 2024-08-30 2dd57fc
update pdoc README badge
