### 2024-09-10 0be5858
Chamber example notebook from barrel study (Tian et al 2017) (#372)
Co-authored-by: Sylwester Arabas <sylwester.arabas@agh.edu.pl>
