### 2024-09-25 612059a
merge-multiple and pattern options for artifact actions
