+++ .github/workflows/readme_listings.yml
@@ -38,7 +38,7 @@ jobs:
       - run: julia readme.jl > readme_output/julia.txt
       - uses: actions/upload-artifact@v4
         with:
-          name: readme_output
+          name: readme_output-julia
           path: readme_output
 
   python:
@@ -65,7 +65,7 @@ jobs:
       - run: python -We readme.py > readme_output/python.txt
       - uses: actions/upload-artifact@v4
         with:
-          name: readme_output
+          name: readme_output-python
           path: readme_output
       
   fortran:
@@ -92,7 +92,7 @@ jobs:
 
       - uses: actions/upload-artifact@v4
         with:
-          name: readme_output
+          name: readme_output-fortran
           path: readme_output
 
   assert:
@@ -103,7 +103,8 @@ jobs:
     - run: pip install numpy
     - uses: actions/download-artifact@v4
       with:
-          name: readme_output
+          pattern: readme_output-*
+          merge-multiple: true
           path: readme_output
     - run : python -c 'import numpy as np; import os; dir="readme_output/"; data=[float(np.loadtxt(dir+file)) for file in os.listdir(dir)]; print("data:", data); similar_as_first = np.array([abs(data[0]-k)/data[0] for k in data[1:]]); print("similar_as_first", similar_as_first); assert((similar_as_first < .5).all())'
 
@@ -150,6 +151,6 @@ jobs:
 
       - uses: actions/upload-artifact@v4
         with:
-          name: readme_output
+          name: readme_output-matlab
           path: readme_output
 
