### 2024-09-26 8ba81ac
fix yaml syntax broken in the last marged PR
