### 2024-09-26 758a1a4
workaround case [in]sensivity issue in sdist filenames blocking PyPI uploads
