+++ .github/workflows/tests+pypi.yml
@@ -179,6 +179,11 @@ jobs:
       - if: matrix.platform == 'ubuntu-latest'
         run: rm dist/*-linux_*
 
+      # case sensitivity issue - avoid coexisting PyPartMC-... and pypartmc-... files
+      # https://github.com/pypi/warehouse/issues/15824
+      - if: matrix.platform == 'windows-latest'
+        run: rm dist/pypartmc*.tar.gz
+
       - uses: actions/upload-artifact@v4
         with:
           name: dist-${{matrix.platform}}-${{matrix.manylinux}}-${{matrix.python-version}}
