### 2024-09-26 9770f02
one more try at fixing PyPI upload
