### 2024-10-20 #179 @cmarmo
Hello @slayoo, I'm going to take care of PyPartMC as editor.
I have started to look for reviewers let's cross fingers ... :crossed_fingers:
