### 2024-10-20 6aed2a8
cleanup leftover files. closes #373
