+++ examples/chamber_data/0925/ref_aero_time.txt
@@ -1,49 +1,0 @@
-0	1.30898E+11	3.25738E-07
-420	1.19515E+11	3.21076E-07
-840	1.09495E+11	3.18544E-07
-1260	1.01068E+11	3.1369E-07
-1680	94134900000	3.1467E-07
-2100	87897200000	3.10829E-07
-2520	82551400000	3.08678E-07
-2940	77647400000	3.04676E-07
-3360	73623800000	3.01067E-07
-3780	67930500000	2.95461E-07
-4200	65888100000	2.90071E-07
-4620	62582400000	2.84073E-07
-5040	59765600000	2.80337E-07
-5460	57719900000	2.8295E-07
-5880	54980200000	2.83708E-07
-6300	52876900000	2.78871E-07
-6720	51163400000	2.76975E-07
-7140	49071700000	2.75866E-07
-7560	46814100000	2.71413E-07
-7980	45264200000	2.67989E-07
-8400	43484200000	2.6377E-07
-8820	42400400000	2.64128E-07
-9240	40871900000	2.58219E-07
-9660	39407200000	2.53514E-07
-10080	38229100000	2.52278E-07
-10500	37014700000	2.4769E-07
-10920	35644000000	2.42154E-07
-11340	34410900000	2.40518E-07
-11760	33681200000	2.43037E-07
-12180	32030300000	2.40009E-07
-12600	31212500000	2.38165E-07
-13020	30291900000	2.34896E-07
-13440	29387600000	2.30355E-07
-13860	28567600000	2.30407E-07
-14280	27855500000	2.28884E-07
-14700	26997800000	2.23884E-07
-15120	26348700000	2.21466E-07
-15540	25458700000	2.15623E-07
-15960	24840600000	2.14156E-07
-16380	24219400000	2.09616E-07
-16800	23710900000	2.05986E-07
-17220	23038300000	2.05518E-07
-17640	22625600000	2.02444E-07
-18060	22069200000	2.03548E-07
-18480	21447500000	1.99564E-07
-18900	20930500000	1.98804E-07
-19320	20511200000	1.98153E-07
-19740	20020700000	1.95232E-07
-20160	19622600000	1.90199E-07
