### 2024-10-20 a89c2a6
cleanup leftover files. closes #373 (#377)
