+++ examples/chamber_data/0925/ref_aero_size_num.txt
@@ -1,115 +1,0 @@
-1.36E-08	6.60E+08	6.53E+07	1.50E+08	2.61E+08	6.53E+07	6.53E+07	1.96E+08	9.09E+07	1.31E+08	1.31E+08	6.53E+07	0	6.53E+07	6.53E+07	0	1.96E+08	6.53E+07	1.31E+08	6.53E+07	0	0	1.96E+08	0	6.58E+06	0	0	0	1.96E+08	1.31E+08	0	1.31E+08	0	6.53E+07	6.53E+07	6.53E+07	0	0	0	0	0	0	0	0	0	0	0	0	0	0
-1.41E-08	6.76E+08	4.34E+08	1.24E+08	6.21E+07	0	6.21E+07	1.24E+08	5.58E+07	0	0	0	0	0	1.24E+08	6.21E+07	1.24E+08	6.21E+07	0	0	0	0	0	6.21E+07	5.58E+07	0	0	0	6.21E+07	6.21E+07	0	0	0	0	6.21E+07	0	0	0	0	6.21E+07	0	2.94E+07	6.21E+07	6.21E+07	6.21E+07	0	6.21E+07	0	0	6.21E+07
-1.46E-08	2.95E+08	1.18E+08	1.77E+08	4.87E+07	0	5.90E+07	5.90E+07	1.77E+08	0	5.90E+07	0	0	0	5.90E+07	5.90E+07	1.18E+08	0	5.90E+07	0	0	0	0	0	0	4.87E+07	0	0	0	0	0	0	0	0	0	0	5.90E+07	0	0	5.90E+07	0	1.49E+08	5.90E+07	0	5.90E+07	0	0	5.90E+07	0	0
-1.51E-08	2.82E+08	3.38E+08	2.82E+08	9.82E+06	5.63E+07	0	8.81E+06	0	1.13E+08	0	0	0	0	0	0	0	1.13E+08	0	0	0	0	1.13E+08	5.63E+07	0	9.82E+06	0	0	5.63E+07	0	0	0	0	0	0	0	0	0	5.63E+07	0	0	0	0	0	0	0	0	0	0	0
-1.57E-08	4.83E+08	3.76E+08	1.32E+08	5.37E+07	5.37E+07	1.07E+08	9.90E+07	0	1.07E+08	1.07E+08	0	0	0	5.37E+07	0	0	0	0	0	0	5.37E+07	0	0	0	0	0	0	5.37E+07	0	0	0	0	0	0	0	0	0	0	0	0	0	2.50E+07	0	0	5.37E+07	0	0	0	0
-1.63E-08	5.12E+08	3.59E+08	1.30E+08	5.12E+07	5.12E+07	0	1.54E+08	5.12E+07	0	0	0	0	0	5.12E+07	0	0	0	0	0	0	0	0	5.12E+07	0	5.12E+07	0	0	5.12E+07	0	0	0	0	0	0	0	0	0	0	0	0	5.12E+07	7.87E+07	5.12E+07	0	0	5.12E+07	0	0	0
-1.68E-08	2.45E+08	2.94E+08	1.47E+08	0	0	1.47E+08	1.47E+08	0	4.89E+07	0	0	0	0	0	9.79E+07	4.89E+07	4.89E+07	4.89E+07	0	4.89E+07	0	0	0	0	0	0	4.89E+07	4.89E+07	4.89E+07	0	0	0	0	0	0	0	0	0	0	0	0	4.89E+07	0	0	0	0	0	0	0
[…]
