### 2024-10-25 fe3536d
enable debug ssh access
