### 2024-10-25 354e86f
bump setup-pytho action and use it on Apple Silicon images
