+++ .github/workflows/tests+pypi.yml
@@ -63,10 +63,6 @@ jobs:
             python-version: "3.10"
           - platform: macos-14
             python-version: "3.11"
-          - platform: macos-14
-            python-version: "3.12"
-          - platform: macos-14
-            python-version: "3.13"
           - manylinux: "manylinux2010_x86_64"
             platform: ubuntu-latest
             python-version: "3.7"
@@ -142,7 +138,7 @@ jobs:
           python -m build 2>&1 | tee build.log
           exit `fgrep -i warning build.log | grep -v "WARNING setuptools_scm" | grep -v "-warnings" | grep -v "All Warnings are enabled" | wc -l`
 
-      - if: startsWith(matrix.platform, 'macos-') && matrix.python-version == '3.11'
+      - if: startsWith(matrix.platform, 'macos-12') && matrix.python-version == '3.11'
         run: |
           # workaround for buggy universal2 wheel name with Python 3.11 (contains x86_64 binary only)
           for whl in dist/*_universal2.whl; do
