### 2024-10-25 d402eff
bump setup-python action and use it on Apple Silicon images (#379)
