+++ .github/workflows/tests+pypi.yml
@@ -60,7 +60,9 @@ jobs:
           - platform: macos-12
             python-version: "3.11"
           - platform: macos-14
-            python-version: "system"
+            python-version: "3.10"
+          - platform: macos-14
+            python-version: "3.11"
           - manylinux: "manylinux2010_x86_64"
             platform: ubuntu-latest
             python-version: "3.7"
@@ -124,18 +126,10 @@ jobs:
           submodules: recursive
           fetch-depth: 0  # https://github.com/pypa/setuptools_scm/issues/480
 
-      - if: matrix.python-version != 'system'
-        uses: actions/setup-python@v4
+      - uses: actions/setup-python@v5.0.0
         with:
           python-version: ${{ matrix.python-version }}
-
-      - if: matrix.python-version == 'system'
-        run: |
-          sudo ln -s `which python3` /usr/local/bin/python
-          echo PIP_INSTALL_OPTS=--break-system-packages >> $GITHUB_ENV
-
-      - if: matrix.python-version != 'system'
-        run: python -m pip install --upgrade pip==22.1.2
+      - run: python -m pip install --upgrade pip==22.1.2
 
       - run: python -m pip install $PIP_INSTALL_OPTS "build<1.1" wheel
       - run: cmake --version
@@ -144,7 +138,7 @@ jobs:
           python -m build 2>&1 | tee build.log
           exit `fgrep -i warning build.log | grep -v "WARNING setuptools_scm" | grep -v "-warnings" | grep -v "All Warnings are enabled" | wc -l`
 
-      - if: startsWith(matrix.platform, 'macos-') && matrix.python-version == '3.11'
+      - if: startsWith(matrix.platform, 'macos-12') && matrix.python-version == '3.11'
         run: |
           # workaround for buggy universal2 wheel name with Python 3.11 (contains x86_64 binary only)
           for whl in dist/*_universal2.whl; do
@@ -219,7 +213,7 @@ jobs:
     runs-on: ubuntu-latest
     needs: [build]
     steps:
-    - uses: actions/setup-python@v2
+    - uses: actions/setup-python@v5.0.0
       with:
         python-version: "3.11"
     - run: pip install twine auditwheel
