+++ .github/workflows/tests+pypi.yml
@@ -138,7 +138,7 @@ jobs:
           python -m build 2>&1 | tee build.log
           exit `fgrep -i warning build.log | grep -v "WARNING setuptools_scm" | grep -v "-warnings" | grep -v "All Warnings are enabled" | wc -l`
 
-      - if: startsWith(matrix.platform, 'macos-')
+      - if: ${{ (matrix.platform == 'macos-12' && matrix.python-version == '3.11') || matrix.platform == 'macos-14' }}
         run: |
           # workaround for buggy universal2 wheel names (contain single arch binary only!)
           for whl in dist/*_universal2.whl; do
