### 2024-10-29 #179 @cmarmo
Hi @slayoo, I am glad to announce that we have our first reviewer! :partying_face:
@simonom, thank you for volunteering to review for pyOpenSci!

I am actively looking for a second reviewer, but I think Simon can start to review, as this submission already waited a bit.
So I'm letting him introduce himself here and continuing my quest.
