### 2024-10-31 46f54f9
add Python 3.12 to list of Linux builds on CI
