### 2024-11-01 d26ec0e
add Windows and macOS 3.12 builds; pin pip to a newer version to fix compatibility with P3.12
