+++ .github/workflows/tests+pypi.yml
@@ -49,6 +49,8 @@ jobs:
             python-version: "3.10"
           - platform: windows-latest
             python-version: "3.11"
+          - platform: windows-latest
+            python-version: "3.12"
           - platform: macos-12
             python-version: "3.7"
           - platform: macos-12
@@ -59,10 +61,14 @@ jobs:
             python-version: "3.10"
           - platform: macos-12
             python-version: "3.11"
+          - platform: macos-12
+            python-version: "3.12"
           - platform: macos-14
             python-version: "3.10"
           - platform: macos-14
             python-version: "3.11"
+          - platform: macos-14
+            python-version: "3.12"
           - manylinux: "manylinux2010_x86_64"
             platform: ubuntu-latest
             python-version: "3.7"
@@ -132,7 +138,7 @@ jobs:
       - uses: actions/setup-python@v5.0.0
         with:
           python-version: ${{ matrix.python-version }}
-      - run: python -m pip install --upgrade pip==22.1.2
+      - run: python -m pip install --upgrade pip==24.3.1
 
       - run: python -m pip install $PIP_INSTALL_OPTS "build<1.1" wheel
       - run: cmake --version
