### 2024-11-01 849052f
drop Python 3.7 from CI builds - it reached EOL and is incompatible with current pip
