+++ .github/workflows/tests+pypi.yml
@@ -40,8 +40,6 @@ jobs:
       matrix:
         include: 
           - platform: windows-latest
-            python-version: "3.7"
-          - platform: windows-latest
             python-version: "3.8"
           - platform: windows-latest
             python-version: "3.9"
@@ -51,8 +49,6 @@ jobs:
             python-version: "3.11"
           - platform: windows-latest
             python-version: "3.12"
-          - platform: macos-12
-            python-version: "3.7"
           - platform: macos-12
             python-version: "3.8"
           - platform: macos-12
@@ -71,9 +67,6 @@ jobs:
             python-version: "3.12"
           - manylinux: "manylinux2010_x86_64"
             platform: ubuntu-latest
-            python-version: "3.7"
-          - manylinux: "manylinux2010_x86_64"
-            platform: ubuntu-latest
             python-version: "3.8"
           - manylinux: "manylinux2010_x86_64"
             platform: ubuntu-latest
@@ -81,9 +74,6 @@ jobs:
           - manylinux: "manylinux2010_x86_64"
             platform: ubuntu-latest
             python-version: "3.10"
-          - manylinux: "manylinux_2_24_x86_64"
-            platform: ubuntu-latest
-            python-version: "3.7"
           - manylinux: "manylinux_2_24_x86_64"
             platform: ubuntu-latest
             python-version: "3.8"
@@ -206,8 +196,7 @@ jobs:
           python -m pytest --durations=10 -v -s -We -p no:unraisableexception .
           cd ..
 
-      - if: matrix.python-version != '3.7'
-        run: |
+      - run: |
           python -m pip install $PIP_INSTALL_OPTS -r gitmodules/devops_tests/requirements.txt
           ex -sc 'g/^PyPartMC/d' -cx .binder/requirements.txt
           python -m pip install $PIP_INSTALL_OPTS -r .binder/requirements.txt
