### 2024-11-01 dfb9a7e
try installint P3.12 using yum on manylinux image
