### 2024-11-02 #179 @Kai-Striega
Hello @slayoo,

@cmarmo commented on the scientific-python discord, asking for reviewers. I'm happy to review this if no one else becomes available soon.

I've done significant work using C++/Fortran + Python. However, I haven't done much with pybind11 or CMake. Happy to put some work in and brush up on those.

Before committing, I’ve signed up to a couple of different things that I’ll need to prioritise. It may take me a while (at the latest, over the Christmas break) to get a quality review done. Is all that acceptable for you? Happy to be superseded if someone else becomes available.
