### 2024-11-05 e8083de
try putting CampCore ctors first?
