+++ src/pypartmc.cpp
@@ -82,8 +82,8 @@ PYBIND11_MODULE(_PyPartMC, m) {
              same, but without the \c _a suffix.
         )pbdoc"
     )
-        .def(py::init<const nlohmann::json&>())
         .def(py::init<const CampCore&>())
+        .def(py::init<const nlohmann::json&>())
         .def("spec_by_name", AeroData::spec_by_name,
              "Returns the number of the species in AeroData with the given name")
         .def("__len__", AeroData::__len__, "Number of aerosol species")
@@ -288,8 +288,8 @@ PYBIND11_MODULE(_PyPartMC, m) {
             is gas_state%%mix_rat(i).
         )pbdoc"
     )
-        .def(py::init<const py::tuple&>())
         .def(py::init<const CampCore&>())
+        .def(py::init<const py::tuple&>())
         .def("__len__", GasData::__len__,
             "returns number of gas species")
         .def_property_readonly("n_spec", GasData::__len__)
