+++ src/pypartmc.cpp
@@ -82,7 +82,7 @@ PYBIND11_MODULE(_PyPartMC, m) {
              same, but without the \c _a suffix.
         )pbdoc"
     )
-        .def(py::init<const CampCore&>())
+        .def(py::init<const nlohmann::json&, const CampCore&>())
         .def(py::init<const nlohmann::json&>())
         .def("spec_by_name", AeroData::spec_by_name,
              "Returns the number of the species in AeroData with the given name")
