### 2024-11-08 e43f0c4
bump CI mac jobs to macos-13 from macos-12 (will be phased out in Dec 24) (#382)
