### 2024-11-08 65438de
disable attestations for gh-action-pypi-publish + pin action version
