### 2024-11-08 3fa5ed8
add basic test for Photolysis ctor
