### 2024-11-26 #179 @slayoo
@simonom, thank you!

So far, we have not yet produced binary wheels for Python 3.13, please stay tuned, hope to roll these out in the coming days at least for macOS.

Since there are no binary wheels available for your platform/OS setup, pip tries to compile the package from source. The failure in getting it compiled seems to originate from an incompatibility around the `setuptools_scm` package which we use to automatically generate PyPartMC version strings, will investigate it as well.

Thanks!
