### 2024-12-02 #179 @cmarmo
Hi @slayoo, I am glad to announce that we have our second reviewer! :partying_face:
@dbaston, thank you for volunteering to review for pyOpenSci!

2. Here is the **[review template](https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html#peer-review-template)** that you will need to fill out and submit here as a comment, once your review is complete.

Now that we have the two reviewers assigned I am tentatively setting a deadline for the review of PyPartMC: let's try to give a nice Christmas gift to PyPartMC maintainers and let the review in for December 23th! 🎄

Thank you all for your work and your patience!
