### 2024-12-05 a768d8d
add Python 3.13 to macOS and Windows builds, remove 3.12 from manylinux
