+++ .github/workflows/tests+pypi.yml
@@ -49,6 +49,8 @@ jobs:
             python-version: "3.11"
           - platform: windows-latest
             python-version: "3.12"
+          - platform: windows-latest
+            python-version: "3.13"
           - platform: macos-12
             python-version: "3.8"
           - platform: macos-12
@@ -59,12 +61,16 @@ jobs:
             python-version: "3.11"
           - platform: macos-12
             python-version: "3.12"
+          - platform: macos-12
+            python-version: "3.13"
           - platform: macos-14
             python-version: "3.10"
           - platform: macos-14
             python-version: "3.11"
           - platform: macos-14
             python-version: "3.12"
+          - platform: macos-14
+            python-version: "3.13"
           - manylinux: "manylinux2010_x86_64"
             platform: ubuntu-latest
             python-version: "3.8"
@@ -86,9 +92,6 @@ jobs:
           - manylinux: "manylinux_2_24_x86_64"
             platform: ubuntu-latest
             python-version: "3.11"
-          - manylinux: "manylinux_2_24_x86_64"
-            platform: ubuntu-latest
-            python-version: "3.12"
       fail-fast: false
     runs-on: ${{ matrix.platform }}
     steps:
@@ -169,7 +172,6 @@ jobs:
           python-versions: ${{ env.PV }}
           build-requirements: 'setuptools_scm'
           pre-build-command: 'git config --global --add safe.directory "*"'
-          system-packages: 'python3.12'
 
       - if: matrix.platform == 'ubuntu-latest'
         run: rm dist/*-linux_*
