### 2024-12-05 00c6670
remove P13 on Windows, fix macos vm names
