+++ .github/workflows/tests+pypi.yml
@@ -49,8 +49,6 @@ jobs:
             python-version: "3.11"
           - platform: windows-latest
             python-version: "3.12"
-          - platform: windows-latest
-            python-version: "3.13"
           - platform: macos-13
             python-version: "3.8"
           - platform: macos-13
@@ -59,9 +57,9 @@ jobs:
             python-version: "3.10"
           - platform: macos-13
             python-version: "3.11"
-          - platform: macos-12
+          - platform: macos-13
             python-version: "3.12"
-          - platform: macos-12
+          - platform: macos-13
             python-version: "3.13"
           - platform: macos-14
             python-version: "3.10"
