### 2024-12-05 586126a
remove P3.13 (due to PyMieScatt incompatibility)
