+++ .github/workflows/tests+pypi.yml
@@ -59,16 +59,12 @@ jobs:
             python-version: "3.11"
           - platform: macos-13
             python-version: "3.12"
-          - platform: macos-13
-            python-version: "3.13"
           - platform: macos-14
             python-version: "3.10"
           - platform: macos-14
             python-version: "3.11"
           - platform: macos-14
             python-version: "3.12"
-          - platform: macos-14
-            python-version: "3.13"
           - manylinux: "manylinux2010_x86_64"
             platform: ubuntu-latest
             python-version: "3.8"
