### 2024-12-06 ce9c623
expose additive kernel coefficient property in Python
