+++ src/env_state.F90
@@ -94,6 +94,28 @@ module PyPartMC_env_state
 
     end subroutine
 
+    subroutine f_env_state_set_additive_kernel_coefficient(ptr_c, additive_kernel_coefficient) bind(C)
+        type(env_state_t), pointer :: ptr_f => null()
+        type(c_ptr), intent(in) :: ptr_c
+        real(c_double), intent(in) :: additive_kernel_coefficient
+
+        call c_f_pointer(ptr_c, ptr_f)
+
+        ptr_f%additive_kernel_coefficient = additive_kernel_coefficient
+
+    end subroutine
+
+    subroutine f_env_state_get_additive_kernel_coefficient(ptr_c, additive_kernel_coefficient) bind(C)
+        type(env_state_t), pointer :: ptr_f => null()
+        type(c_ptr), intent(in) :: ptr_c
+        real(c_double), intent(out) :: additive_kernel_coefficient
+        
+        call c_f_pointer(ptr_c, ptr_f)
+
+        additive_kernel_coefficient = ptr_f%additive_kernel_coefficient
+
+    end subroutine
+
     subroutine f_env_state_set_pressure(ptr_c, pressure) bind(C)
         type(env_state_t), pointer :: ptr_f => null()
         type(c_ptr), intent(in) :: ptr_c
