### 2024-12-06 2e48fc6
add Python 3.12 to list of macOS and Windows CI builds; drop Python 3.7 (#381)
