+++ .github/workflows/tests+pypi.yml
@@ -40,8 +40,6 @@ jobs:
       matrix:
         include: 
           - platform: windows-latest
-            python-version: "3.7"
-          - platform: windows-latest
             python-version: "3.8"
           - platform: windows-latest
             python-version: "3.9"
@@ -49,8 +47,8 @@ jobs:
             python-version: "3.10"
           - platform: windows-latest
             python-version: "3.11"
-          - platform: macos-13
-            python-version: "3.7"
+          - platform: windows-latest
+            python-version: "3.12"
           - platform: macos-13
             python-version: "3.8"
           - platform: macos-13
@@ -59,13 +57,14 @@ jobs:
             python-version: "3.10"
           - platform: macos-13
             python-version: "3.11"
+          - platform: macos-13
+            python-version: "3.12"
           - platform: macos-14
             python-version: "3.10"
           - platform: macos-14
             python-version: "3.11"
-          - manylinux: "manylinux2010_x86_64"
-            platform: ubuntu-latest
-            python-version: "3.7"
+          - platform: macos-14
+            python-version: "3.12"
           - manylinux: "manylinux2010_x86_64"
             platform: ubuntu-latest
             python-version: "3.8"
@@ -75,9 +74,6 @@ jobs:
           - manylinux: "manylinux2010_x86_64"
             platform: ubuntu-latest
             python-version: "3.10"
-          - manylinux: "manylinux_2_24_x86_64"
-            platform: ubuntu-latest
-            python-version: "3.7"
           - manylinux: "manylinux_2_24_x86_64"
             platform: ubuntu-latest
             python-version: "3.8"
@@ -129,7 +125,7 @@ jobs:
       - uses: actions/setup-python@v5.0.0
         with:
           python-version: ${{ matrix.python-version }}
-      - run: python -m pip install --upgrade pip==22.1.2
+      - run: python -m pip install --upgrade pip==24.3.1
 
       - run: python -m pip install $PIP_INSTALL_OPTS "build<1.1" wheel
       - run: cmake --version
@@ -197,8 +193,7 @@ jobs:
           python -m pytest --durations=10 -v -s -We -p no:unraisableexception .
           cd ..
 
-      - if: matrix.python-version != '3.7'
-        run: |
+      - run: |
           python -m pip install $PIP_INSTALL_OPTS -r gitmodules/devops_tests/requirements.txt
           ex -sc 'g/^PyPartMC/d' -cx .binder/requirements.txt
           python -m pip install $PIP_INSTALL_OPTS -r .binder/requirements.txt
