### 2024-12-06 6b412ca
bump PartMC submodule and expose additive kernel coefficient property (#385)
