### 2024-12-06 463737a
use PyMieScatt from repo instead of PyPI; add P3.13 to macOS CI;
