### 2024-12-06 ecc4758
use P3.13-compatible Numba with Binder
