+++ .binder/requirements.txt
@@ -3,6 +3,7 @@ matplotlib
 ipywidgets
 voila
 open-atmos-jupyter-utils
+numba==0.61.0rc1
 PySDM
 git+https://github.com/bsumlin/PyMieScatt.git
 SciPy
