### 2024-12-06 2b0bd7a
use env marker for Numba req
