+++ .binder/requirements.txt
@@ -3,7 +3,7 @@ matplotlib
 ipywidgets
 voila
 open-atmos-jupyter-utils
-numba==0.61.0rc1
+numba>=0.61.0rc1; python_version >= '3.13'
 PySDM
 git+https://github.com/bsumlin/PyMieScatt.git
 SciPy
