### 2024-12-07 79df5b9
use PyMieScatt from repo instead of PyPI; add P3.13 to macOS CI (#386)
