+++ .github/workflows/tests+pypi.yml
@@ -59,6 +59,8 @@ jobs:
             python-version: "3.11"
           - platform: macos-13
             python-version: "3.12"
+          - platform: macos-13
+            python-version: "3.13"
           - platform: macos-14
             python-version: "3.10"
           - platform: macos-14
