### 2024-12-07 bd3e22c
enable macos-14/P3.13 builds on CI with temporarily disabled notebook checks for this combination (due to Numba incompatibility) (#388)
