### 2024-12-07 #179 @slayoo
@simonom

>> I am struggling with `pip install PyPartMC` at command line on an Apple M2 Pro (macOS Sonoma).

> So far, we have not yet produced binary wheels for Python 3.13, please stay tuned, hope to roll these out in the coming days at least for macOS.

As of PyPartMC v1.3.10 (released earlier today), there are binary wheels on PyPI for Python 3.13 on macOS. Supported OS versions are: macosx-13 (universal binary) and macosx-14 (ARM binary). Hope it helps! Feedback welcome.

Note that Python 3.13 is not fully supported in auxiliary packages used in PyPartMC example notebooks, notably:
- PyMieScatt used in the [mie_optical.ipynb](https://github.com/open-atmos/PyPartMC/blob/main/examples/mie_optical.ipynb) example requires installation from GitHub instead of PyPI (`pip install git+https://github.com/bsumlin/PyMieScatt.git`) to work due to Python 3.13 implying newer SciPy ([reported upstream](https://github.com/bsumlin/PyMieScatt/pull/26#issuecomment-2521677787))
- PySDM used in [lognorm_ex.ipynb](https://github.com/open-atmos/PyPartMC/blob/main/examples/lognorm_ex.ipynb) and in [terminal_velocities.ipynb](https://github.com/open-atmos/PyPartMC/blob/main/examples/terminal_velocities.ipynb) uses Numba which requires pre-release version to be compatible with Python 3.13 (`pip install numba==0.61.0rc1`), but it anyhow fails on ARM CPUs ([reported upstream](https://github.com/numba/numba/issues/9413#issuecomment-2524697307))
