### 2024-12-09 1d85c54
bump pybind11_json to 0.2.15 and remove one test skip given a segfault was fixed. closes #319
