### 2024-12-16 #179 @dbaston
I did not find instructions for installing the development version. After pulling the submodules, running `pip install .` from the project root successfully installed the package.

I ran the notebooks on Google Colab without issue. It would be helpful to have instructions for running the notebooks locally, as additional dependencies such as `open_atmos_jupyter_utils` are required. The notebooks would benefit from narrative text explaining the purpose of each notebook, and an explanation of what is being performed in each cell.

While the documentation contains an entry for user-facing functions, it isn't always clear what the expected arguments or return values are. Some argument names are not rendered in the documentation. For example, `Scenario.init_env_state` lists argument names `arg0` and `arg1` instead of the names `env_state` and `time` that can be found in the sources. In other cases, argument types indicate that JSON is expected but do not describe the expected content of that JSON. Some functions also lack descriptions (e.g., `AeroMode.vol_frac`, `AeroMode.gsd`). The markup language used in the docstings is not fully recognized when generating HTML documentation and produces output such as `\c name(i) is the name of that species`.

I did not find examples in the Python API documentation.

It would be helpful to describe how someone can contribute to the documentation. This may be as simple as describing where the docstring text can be found.

Included in `setup.py` file. The project does not have a `pyproject.toml`.

None required.

Described in the linked paper.

Did not evaluate.

Did not evaluate.

Did not see instructions for running tests. Tests run successfully using `pytest tests`.

Did not evaluate. A code coverage report of the C++ binding code would be useful.

PyPartMC uses a custom `setup.py` instead of the `pyproject.toml` discussed in the packaging guidelines. Because the project is built using CMake, it may be straightforward to switch to a `pyproject.toml` and `scikit-build-core`.

It wasn't clear to me why C++ class methods were implemted as static methods manually taking a reference to `self`. The pattern is followed consistently, however.

---

This package cleanly exposes Fortan code to Python using pybind11. I did not see any serious code quality issues. Tests run smoothly but I was not able to easily determine which code is covered by the tests.

HDF and netCDF being fairly common libraries, I wonder about the potential for runtime issues if the version vendored by PyPartMC differs from a version loaded by another package such as `netCDF4` or `h5py`.
