### 2024-12-20 #179 @simonom
The authors could  be more explicit in the README introduction, such as providing an example of the user you are serving. Actually, the goals stated in the ‘Target audience and scientific applications’ sections of the review page ([PyPartMC · Issue #179 · pyOpenSci/software-submission](https://github.com/pyOpenSci/software-submission/issues/179)) seem very suitable for this section.

The user could benefit from some description around the pip install PyPartMC command to explain that CMake and a fortran compiler are required. This is well detailed in the FAQs, but it makes sense for these dependencies to be stated before the pip install instruction for PyPartMC, including the necessary versions (e.g. python).

The API documentation is very valuable for explaining how a user can set inputs for functions, but the link is somewhat hidden in the Features section of README. I suggest a more obvious presence in README.

In CITATION.cff is author information in addition to a doi to an associated published paper. In Credits of README is funding information.

- [x ] Package has a README.md file in the root directory.

Would be useful to have a repostatus badge.

The authors could  be more explicit in the README introduction, such as providing an example of the user you are serving. Actually, the goals stated in the ‘Target audience and scientific applications’ sections of the review page ([PyPartMC · Issue #179 · pyOpenSci/software-submission](https://github.com/pyOpenSci/software-submission/issues/179)) seem very suitable for this section.

Because the README does not contain a section titled ‘Installation’ it is really easy to miss the ‘pip install PyPartMC’ command, so recommend making this clearer. Preferably this installations section would also describe the dependencies on Fortran and CMake.

Not applicable for local install, and additional setup for vignettes is well described

I think the vignettes are a real strong point of the repository.

The python and Julia example are provided. Like with the Short description of package goals (for which I recommend an example of a target user), it would help to have more detail in the ‘Usage examples’ introduction about why a user would be interested in randomly sampling particles from an aerosol size distribution. I think this is particularly important because PyPartMC is aiming to draw in more novice users of PartMC, so these people will potentially be new to PartMC.

As stated by me in the  **Function Documentation:** section above, it would be useful to have the link to the API docs more visible than the brief mention in the Features section.

I think this section needs including in the README, particularly as the package is an interface to another package (PartMC). This section would be an opportunity to really clearly explain what the implications of using this package are over using the PartMC package directly. It would also be good to include in this section the references included in ‘Other Python packages with relevant feature scope’ in the review page ([PyPartMC · Issue #179 · pyOpenSci/software-submission](https://github.com/pyOpenSci/software-submission/issues/179))

I think this section needs including in the README, particularly as the package is an interface to another package (PartMC). This section would be an opportunity to really clearly explain what the implications of using this package are over using the PartMC package directly. It would also be good to include in this section the references included in ‘Other Python packages with relevant feature scope’ in the review page ([PyPartMC · Issue #179 · pyOpenSci/software-submission](https://github.com/pyOpenSci/software-submission/issues/179))

See notes above on clearer linking to the API Docs in README

See notes above on clearer linking to the API Docs in README

The API Docs are impressive, and some examples are also provided

Some tests need pytest installed, so this should be stated along with other install depenencies

I can see that CI is setup in the .github/workflows folder, but note that the link in the ‘CI builds’ text in ‘Features’ of README is broken

---

I've enjoyed reviewing this package, which has been very well put together overall. The comments I have provided I believe cover just minor changes.
