+++ .github/workflows/tests+pypi.yml
@@ -195,6 +195,10 @@ jobs:
           for i in dist/*.whl; do python -m pip install $PIP_INSTALL_OPTS $i[tests]; done;
           python -m pip show --verbose PyPartMC
 
+      - uses: julia-actions/setup-julia@v2
+        with:
+          version: '1.11'
+    
       ### make sure PyPartMC is not picked up from a local folder
       - run: |
           cd tests
