### 2024-12-23 88bc802
use "no-local-version" (instead of a lambda) for setuptools_scm "local_scheme"; avoid warning-triggering matplotlin 3.10.0 release on CI (#391)
