### 2024-12-30 6541f7b
enable debug ssh access
