### 2025-01-03 5435793
Add: Basic InputGuard implementation
