### 2025-01-03 3dc6951
Add: Integrated InputGuard into the library's basic inner workflow
