### 2025-01-11 #179 @cmarmo
Hello everybody and happy new year to all of you!
I hope you enjoyed some holiday break, we are now back to business and I wondering if we can set a timeline for the answers to the reviewers.

@slayoo, do you think it will be possible for you to answer the comments and fix the fixable issues by the end of January?
Thank you very much for your work.
