### 2025-01-20 9fcf0b2
Add: InputGuard now processes reading of integers and arrays
