### 2025-01-21 99d0acf
Add&fix: InputGuard now catches boolean values. Fixed typo in TestCondense::test_equilib_particle
