### 2025-01-21 99a8525
Fix: Deleted 'do_optical' from an initializing loop in RunPartOpt object. Deleted 'do_optical' from RUN_PART_OPT_CTOR_ARG_SIMULATION test parameter config. All pytest tests are passing
