### 2025-01-28 fce350d
ReadME buttons, flag for Adaptive
