### 2025-01-28 05cd3b2
README update with pyOpenSci review suggestions: repotatus badge; target-audience paragraph; Fortran/CMake dependency mentioned earlier on
