### 2025-01-28 2c05b68
setup codecov for C++ and Fortran files in PyPartMC src folder
