### 2025-01-29 e305720
pin to older Ubuntu to avoid newer-GCC/lcov clash
