+++ .github/workflows/codecov.yml
@@ -8,7 +8,7 @@ on:
 
 jobs:
   codecov:
-    runs-on: ubuntu-latest
+    runs-on: ubuntu-22.04
     steps:
       - uses: actions/checkout@v4.2.2
         with:
