### 2025-01-29 eae74d7
simplify codecov workflow
