+++ .github/workflows/codecov.yml
@@ -16,10 +16,8 @@ jobs:
           fetch-depth: 0
       - run: sudo apt install -y lcov
       - run: |
+          set -xe
           CMAKE_ARGS="-DCMAKE_CXX_FLAGS=--coverage -DCMAKE_C_FLAGS=--coverage -DCMAKE_Fortran_FLAGS=--coverage" pip install -e .[tests]
-          lcov --capture --include="src/*" --initial --directory . --output-file coverage1.info
           pytest -We tests
-          lcov --capture --include="src/*" --directory . --output-file coverage2.info
-          lcov --add-tracefile ./coverage1.info --add-tracefile ./coverage2.info --output-file coveragefused.info
-          lcov -r coveragefused.info "/usr*" -o coveragefused.info
+          lcov --all --capture --directory CMakeFiles/_PyPartMC.dir/src/ --output-file coverage.info
       - uses: codecov/codecov-action@v5.3.1
