### 2025-01-29 26e7ecd
try with src dir spec
