### 2025-01-29 d5d64d8
enable debug access
