+++ .github/workflows/codecov.yml
@@ -19,10 +19,10 @@ jobs:
           set -xe
           CMAKE_ARGS="-DCMAKE_CXX_FLAGS=--coverage -DCMAKE_C_FLAGS=--coverage -DCMAKE_Fortran_FLAGS=--coverage" pip install -e .[tests]
           pytest -We tests
-          lcov --capture --directory src --output-file coverage.info --no-external
-      - uses: mxschmitt/action-tmate@v3
-        with:
-          limit-access-to-actor: true
+          lcov --capture --directory ./build/*/_PyPartMC/CMakeFiles/_PyPartMC.dir/ --output-file coverage.info --no-external
+      # - uses: mxschmitt/action-tmate@v3
+      #   with:
+      #     limit-access-to-actor: true
       - uses: codecov/codecov-action@v5.3.1
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
