+++ .github/workflows/codecov.yml
@@ -26,3 +26,6 @@ jobs:
       - uses: codecov/codecov-action@v5.3.1
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
+          disable_search: true
+          fail_ci_if_error: true
+          files: coverage.info
