### 2025-01-29 d4043f1
fix find/delete
