### 2025-01-29 ce36220
setup codecov for C++ and Fortran files in PyPartMC src folder (#397)
