### 2025-01-30 5d43269
add which julia logic, point readme badges to main
