### 2025-01-30 51abd46
skip Julia setup for CI workflow (which will also disable the Julia runs in the notebook)
