### 2025-01-30 1fda4b3
clarify dependencies sentence
