### 2025-01-30 ef4b69f
README update with pyOpenSci review suggestions: repotatus badge; target-audience paragraph; Fortran/CMake dependency mentioned earlier on (#396)
