### 2025-01-31 79e598b
re-enable tests for macOS 14 & Python 3.13
