+++ .github/workflows/tests+pypi.yml
@@ -198,8 +198,7 @@ jobs:
           python -m pytest --durations=10 -v -s -We -p no:unraisableexception .
           cd ..
 
-      - if: ${{ !(matrix.platform == 'macos-14' && matrix.python-version == '3.13') }}
-        run: |
+      - run: |
           python -m pip install $PIP_INSTALL_OPTS -r gitmodules/devops_tests/requirements.txt
           ex -sc 'g/^PyPartMC/d' -cx .binder/requirements.txt
           python -m pip install $PIP_INSTALL_OPTS -r .binder/requirements.txt
