+++ .github/workflows/tests+pypi.yml
@@ -191,8 +191,7 @@ jobs:
           for i in dist/*.whl; do python -m pip install $PIP_INSTALL_OPTS $i[tests]; done;
           python -m pip show --verbose PyPartMC
 
-      - if: matrix.platform != 'windows-latest'
-        uses: julia-actions/setup-julia@v2
+      - uses: julia-actions/setup-julia@v2
         with:
           version: '1.11'
     
@@ -210,11 +209,12 @@ jobs:
           python -m pip install $PIP_INSTALL_OPTS -r .binder/requirements.txt
           GITHUB_TOKEN=${{ secrets.GITHUB_TOKEN }} python -m pytest --durations=10 -v -s -We -p no:unraisableexception gitmodules/devops_tests
 
-### uncomment to gain ssh access in case of failure
+## uncomment to gain ssh access in case of failure
+      - if: ${{ !(matrix.platform == 'macos-14' && matrix.python-version == '3.12') }}
       # - if: ${{ failure() }}
-      #   uses: mxschmitt/action-tmate@v3
-      #   with:
-      #     limit-access-to-actor: true
+        uses: mxschmitt/action-tmate@v3
+        with:
+          limit-access-to-actor: true
 
   dist_check:
     runs-on: ubuntu-latest
