+++ .github/workflows/tests+pypi.yml
@@ -210,11 +210,10 @@ jobs:
           GITHUB_TOKEN=${{ secrets.GITHUB_TOKEN }} python -m pytest --durations=10 -v -s -We -p no:unraisableexception gitmodules/devops_tests
 
 ## uncomment to gain ssh access in case of failure
-      - if: ${{ !(matrix.platform == 'macos-14' && matrix.python-version == '3.12') }}
       # - if: ${{ failure() }}
-        uses: mxschmitt/action-tmate@v3
-        with:
-          limit-access-to-actor: true
+      #   uses: mxschmitt/action-tmate@v3
+      #   with:
+      #     limit-access-to-actor: true
 
   dist_check:
     runs-on: ubuntu-latest
