### 2025-01-31 0908345
Addition of additive coagulation example including comparison against PySDM and Droplets.jl (#376)
Co-authored-by: Jeff Curtis <jcurtis2@illinois.edu>
Co-authored-by: Sylwester Arabas <sylwester.arabas@agh.edu.pl>
Co-authored-by: Sylwester Arabas <sylwester.arabas@uj.edu.pl>
