+++ .github/workflows/tests+pypi.yml
@@ -191,6 +191,10 @@ jobs:
           for i in dist/*.whl; do python -m pip install $PIP_INSTALL_OPTS $i[tests]; done;
           python -m pip show --verbose PyPartMC
 
+      - uses: julia-actions/setup-julia@v2
+        with:
+          version: '1.11'
+    
       ### make sure PyPartMC is not picked up from a local folder
       - run: |
           cd tests
@@ -204,11 +208,11 @@ jobs:
           python -m pip install $PIP_INSTALL_OPTS -r .binder/requirements.txt
           GITHUB_TOKEN=${{ secrets.GITHUB_TOKEN }} python -m pytest --durations=10 -v -s -We -p no:unraisableexception gitmodules/devops_tests
 
-### uncomment to gain ssh access in case of failure
-#      - if: ${{ failure() }}
-#        uses: mxschmitt/action-tmate@v3
-#        with:
-#          limit-access-to-actor: true
+## uncomment to gain ssh access in case of failure
+      # - if: ${{ failure() }}
+      #   uses: mxschmitt/action-tmate@v3
+      #   with:
+      #     limit-access-to-actor: true
 
   dist_check:
     runs-on: ubuntu-latest
