### 2025-01-31 #179 @slayoo
Just confirming we're working towards addressing the feedback from both reviews. A couple of PRs already merged over the last days, including codecov integration (https://github.com/open-atmos/PyPartMC/pull/397), README updates (https://github.com/open-atmos/PyPartMC/pull/396), more narrative in the example notebooks (https://github.com/open-atmos/PyPartMC/pull/398, https://github.com/open-atmos/PyPartMC/pull/383); let me report a point-by-point summary by the end of the weekend, thank you for your patience.
