### 2025-02-03 cd3f603
Add "examples" variant for pip installation including dependencies needed for examples, mention in README, and use in .binder/requirements.txt file. Closes #305
