### 2025-02-03 f9e45cb
fix PyMieScatt spec
