### 2025-02-03 3e5b840
add "other packages with relevant feature scope" section to README.md
