### 2025-02-03 0e9a826
add "Acknnowledgement and citations" section to README.md
