### 2025-02-03 open-atmos/PyPartMC#406 @slayoo
enhance CONTRIBUTING.md file
- [ ] explain how to run the tests
- [ ] explain how to add new tests
- [ ] explain  why in the C++ code we use static methods manually taking a reference to `self`
