### 2025-02-03 #179 @slayoo
Here's a point-by-point reply to the comments by @dbaston. Thank you for very helpful feedback!

> I did not find instructions for installing the development version. After pulling the submodules, running `pip install .` from the project root successfully installed the package.

The CONTRIBUTING.md file features a section with a brief info on how to set up a development (debugging) setup: https://github.com/open-atmos/PyPartMC/blob/main/CONTRIBUTING.md#how-to-debug
A new issue was created to track improvements to the CONTRIBUTING.md file: https://github.com/open-atmos/PyPartMC/issues/406

> I ran the notebooks on Google Colab without issue. It would be helpful to have instructions for running the notebooks locally, as additional dependencies such as `open_atmos_jupyter_utils` are required.

This information was previously hidden in the `.binder/requirements.txt` file (used also in the CI). In https://github.com/open-atmos/PyPartMC/pull/402, we have moved this metadata into `setup.py` introducing the "examples" installation variant. An indication that `pip install PyPartMC[examples]` can be used to obtain dependencies for the examples was added to the README.md.

> The notebooks would benefit from narrative text explaining the purpose of each notebook, and an explanation of what is being performed in each cell.

PR's https://github.com/open-atmos/PyPartMC/pull/398, https://github.com/open-atmos/PyPartMC/pull/383 and https://github.com/open-atmos/PyPartMC/pull/376 introduced narrative text to three of the examples. There is, clearly, still room for improvement, but we expect that these three examples will also serve as templates for further developments of the descriptive part of PyPartMC example notebooks.

> While the documentation contains an entry for user-facing functions, it isn't always clear what the expected arguments or return values are. Some argument names are not rendered in the documentation. For example, `Scenario.init_env_state` lists argument names `arg0` and `arg1` instead of the names `env_state` and `time` that can be found in the sources. In other cases, argument types indicate that JSON is expected but do not describe the expected content of that JSON. Some functions also lack descriptions (e.g., `AeroMode.vol_frac`, `AeroMode.gsd`). The markup language used in the docstings is not fully recognized when generating HTML documentation and produces output such as `\c name(i) is the name of that species`.
> I did not find examples in the Python API documentation.

In short, this is not addressed yet. We have, though, linked the above comments from two outstanding issues relevant to the docs:
- https://github.com/open-atmos/PyPartMC/issues/31(idea to somehow automatically check the consistency of PyPartMC docstrings and PartMC doxygen string, so that these become maintainable; an automatic conversion from the Doxygen hypertext syntax to something properly renderable by pdoc should be part of it)
- https://github.com/open-atmos/PyPartMC/issues/99 (need to fix hyperlinks for PyPartMC datatypes in the pdoc-generated HTML)

> It would be helpful to describe how someone can contribute to the documentation. This may be as simple as describing where the docstring text can be found.

note taken in the CONTRIBUTING.md-enhancements issue: https://github.com/open-atmos/PyPartMC/issues/406

> Included in `setup.py` file. The project does not have a `pyproject.toml`.

For now, a new issue (with a "help wanted" label) was created: https://github.com/open-atmos/PyPartMC/issues/403

> A [repostatus.org](https://www.repostatus.org/) badge

Added in https://github.com/open-atmos/PyPartMC/pull/396

> [ ]  All functions have documentation and associated examples for use

To this end, one idea is to introduce a mechanism that will automatically verify if all docstrings in PyPartMC match the Doxygen comments from PartMC Fortran code. It is not entirely doable (perhaps this could be limitted to a subset of the docstring), but a starting point for brainstorming: https://github.com/open-atmos/PyPartMC/issues/31

> Did not see instructions for running tests. Tests run successfully using `pytest tests`.

The CONTRIBUTING.md file has a section hinting how to run the tests: https://github.com/open-atmos/PyPartMC/blob/main/CONTRIBUTING.md#how-to-debug

> A code coverage report of the C++ binding code would be useful.

Added in https://github.com/open-atmos/PyPartMC/pull/397
It covers both C++ and Fortran code of the bindings.
Reports and stats available at: https://app.codecov.io/gh/open-atmos/PyPartMC

> PyPartMC uses a custom `setup.py` instead of the `pyproject.toml` discussed in the packaging guidelines. Because the project is built using CMake, it may be straightforward to switch to a `pyproject.toml` and `scikit-build-core`.

Not completed yet, but an issue was created to track progress on it: https://github.com/open-atmos/PyPartMC/issues/403

> It wasn't clear to me why C++ class methods were implemted as static methods manually taking a reference to `self`. The pattern is followed consistently, however.

Noted in the CONTRIBUTING.md0enhancements issue to clarify the above pattern: https://github.com/open-atmos/PyPartMC/issues/406

> HDF and netCDF being fairly common libraries, I wonder about the potential for runtime issues if the version vendored by PyPartMC differs from a version loaded by another package such as `netCDF4` or `h5py`.

Earlier in the development, we have noted that SciPy includes parts of SUNDIALS which are also statically linked within PyPartMC. SciPy is used in the examples, so the CI runs include cases where both are loaded (relevant notes: https://github.com/open-atmos/PyPartMC/issues/16). It seems that it is not an issue, but indeed a clarification why would best be added, e.g., to the CONTRIBUTING.md file: note taken in https://github.com/open-atmos/PyPartMC/issues/406
