### 2025-02-03 7239b4f
add a link to docs earlier in the README (#407)
