### 2025-02-03 f6ea829
rename "TL;DR" to "Installation" in README.md
