### 2025-02-03 bbe0226
fix CI status badge in README.md (#410)
