### 2025-02-03 964b20d
Add "examples" variant for pip installation including dependencies needed for examples, mention in README, and use in .binder/requirements.txt file. Closes #305 (#402)
