### 2025-02-03 #179 @slayoo
Here's a point-by-point reply to the comments by @simonom. Thank you for very helpful feedback!

> The authors could be more explicit in the README introduction, such as providing an example of the user you are serving. Actually, the goals stated in the ‘Target audience and scientific applications’ sections of the review page ([PyPartMC · Issue #179 · pyOpenSci/software-submission](https://github.com/pyOpenSci/software-submission/issues/179)) seem very suitable for this section.

Added in https://github.com/open-atmos/PyPartMC/pull/396

> The user could benefit from some description around the pip install PyPartMC command to explain that CMake and a fortran compiler are required. This is well detailed in the FAQs, but it makes sense for these dependencies to be stated before the pip install instruction for PyPartMC, including the necessary versions (e.g. python).

The following passage was added early in the README as a part of https://github.com/open-atmos/PyPartMC/pull/396:
>> Note that, depending on the environment (OS, hardware, Python version), the pip-install invocation may either trigger a download of a pre-compiled binary, or trigger compilation of PyPartMC. In the latter case, a Fortran compiler and some development tools includiong CMake, m4 and perl are required (while all non-Python dependencies are included in the PyPartMC source archive). In both cases, all Python dependencies will be resolved by pip.

> The API documentation is very valuable for explaining how a user can set inputs for functions, but the link is somewhat hidden in the Features section of README. I suggest a more obvious presence in README.

A sentence with a link to the docs was added to the second paragraph of the README in https://github.com/open-atmos/PyPartMC/pull/407

> Would be useful to have a repostatus badge.

Added in https://github.com/open-atmos/PyPartMC/pull/396

> The authors could be more explicit in the README introduction, such as providing an example of the user you are serving. Actually, the goals stated in the ‘Target audience and scientific applications’ sections of the review page ([PyPartMC · Issue #179 · pyOpenSci/software-submission](https://github.com/pyOpenSci/software-submission/issues/179)) seem very suitable for this section.

ditto - added in https://github.com/open-atmos/PyPartMC/pull/396

> Because the README does not contain a section titled ‘Installation’ it is really easy to miss the ‘pip install PyPartMC’ command, so recommend making this clearer. Preferably this installations section would also describe the dependencies on Fortran and CMake.

"Installation" section added in https://github.com/open-atmos/PyPartMC/pull/408

> The python and Julia example are provided. Like with the Short description of package goals (for which I recommend an example of a target user), it would help to have more detail in the ‘Usage examples’ introduction about why a user would be interested in randomly sampling particles from an aerosol size distribution. I think this is particularly important because PyPartMC is aiming to draw in more novice users of PartMC, so these people will potentially be new to PartMC.

With the README growing in size, we opt for focusing on explanations of the concepts such as sampling in the examples. To this end, the "urban plume" example was extended with broader narrative commenting on each step of a simulation: https://github.com/open-atmos/PyPartMC/pull/398

> As stated by me in the **Function Documentation:** section above, it would be useful to have the link to the API docs more visible than the brief mention in the Features section.

Addressed in https://github.com/open-atmos/PyPartMC/pull/407

>> If applicable, how the package compares to other similar packages and/or how it relates to other packages in the scientific ecosystem.
> I think this section needs including in the README, particularly as the package is an interface to another package (PartMC). This section would be an opportunity to really clearly explain what the implications of using this package are over using the PartMC package directly. It would also be good to include in this section the references included in ‘Other Python packages with relevant feature scope’ in the review page ([PyPartMC · Issue #179 · pyOpenSci/software-submission](https://github.com/pyOpenSci/software-submission/issues/179))

List of other packages with relevant features added in https://github.com/open-atmos/PyPartMC/pull/404

>>  Citation information

Added in https://github.com/open-atmos/PyPartMC/pull/405

> Some tests need pytest installed, so this should be stated along with other install depenencies

The CONTRIBUTING.md file is undergoing enhancements (https://github.com/open-atmos/PyPartMC/pull/409) addressing several issues note in the reviews (https://github.com/open-atmos/PyPartMC/issues/406).

> I can see that CI is setup in the .github/workflows folder, but note that the link in the ‘CI builds’ text in ‘Features’ of README is broken

Fixed in https://github.com/open-atmos/PyPartMC/pull/410
