### 2025-02-03 #179 @slayoo
@cmarmo, @dbaston, @simonom
Thank you for your valuable feedback and the effort, which we also acknowledged in the Release notes for the just-released PyPartMC v1.4.0: https://github.com/open-atmos/PyPartMC/releases/tag/v1.4.0
