### 2025-02-03 #179 @slayoo
And thanks to all PyPartMC developers who contributed to the recent sprint: @nriemer, @mwest1066, @jcurtis2, @emmacware, @zdaq12, @klee-48 & @Griger5
