### 2025-02-03 3a377b8
Fix unicode characters in README breaking builds on Windows
