### 2025-02-04 5520dd6
fix PyMieScatt dependency specification (git URL not accepted by PyPI)
