### 2025-02-04 5735894
force reinstallation from github of PyMieScatt on CI
