### 2025-02-04 b7a5109
fix PyMieScatt dependency specification (git URL not accepted by PyPI) (#412)
