### 2025-02-05 #179 @cmarmo
Thank you @slayoo and all the contributors for your work.

@dbaston , @simonom are the maintainer answer acceptable?
If I understand correctly all the comments weren't addressable right now, so issues are still open, but they are acknowledged by the authors who are willing to address them in the future.
Is this kind of approach fine with you?

Please let us know.

As the authors do not plan to submit to JOSS, this is the last step before acceptance.
