### 2025-02-11 0cb419d
Updated: Updated InputGuards inner logic, so it works properly with more compliacted JSONs
