+++ src/json_resource.hpp
@@ -18,7 +18,9 @@
 struct InputGuard {
     public:
         InputGuard(const nlohmann::json &j) {
-            map_inputs_recursive(j);
+            process_json(j);
+
+            dict_key_present = false;
         }
 
         ~InputGuard() {
@@ -26,36 +28,138 @@ struct InputGuard {
         }
 
         void mark_used_input(const std::string &input_name) {
-            this->used_inputs[input_name] = true;
+            if (prefixes.find(input_name) == prefixes.end()) {
+                std::string prefix = combine_str_vec(curr_prefix);
+
+                if (!prefix.empty()) {
+                    used_inputs[prefix + "/" + input_name] = true;
+                }
+                else {
+                    used_inputs[input_name] = true;
+                }
+            }
+        }
+
+        void update_dict_key(std::string dict_key) {
+            curr_dict_key = dict_key;
+        }
+
+        void check_read_line(std::string line) {
+            if (line == "mode_name") {
+                if (dict_key_present) {
+                    curr_prefix.pop_back();
+                }
+                curr_prefix.push_back(curr_dict_key);
+                dict_key_present = true;
+            }
+            else if (line.empty()) {
+                curr_prefix.pop_back();
+                dict_key_present = false;
+            }
+        }
+
+        void open_spec_file(std::string spec_file_name) {
+            curr_prefix.push_back(spec_file_name);
+        }
+
+        void close_spec_file() {
+            curr_prefix.pop_back();
         }
 
     private:
         std::map<std::string, bool> used_inputs;
 
-        void map_inputs_recursive(const nlohmann::json &j) {
-            if (j.is_array()) {
-                for (auto item : j) {
-                    map_inputs_recursive(item);
-                }
-                return;
-            }
+        std::set<std::string> prefixes;
+        std::string curr_dict_key;
+        std::vector<std::string> curr_prefix;
 
-            for(auto [key, value] : j.items())
-            {
-                if (value.is_structured()) {
-                    map_inputs_recursive(value);
-                }
-                
-                if (key.find_first_not_of(" \t\n\v\f\r") != std::string::npos) 
-                    this->used_inputs[key] = false;
-            }
-        }
+        bool dict_key_present;
 
         void check_used_inputs() {
             for (auto item : used_inputs) {
                 if (!item.second) {
                     throw std::logic_error(std::string("Failed: \"") + item.first + std::string("\" parameter remains unused."));
                 }
+            }
+        }
+        
+        void process_json(const nlohmann::json &j) {
+            nlohmann::json flat = j.flatten();
+
+            // JSON Pointer, as in a string syntax for identifying a specific value in JSON
+            std::vector<std::string> json_pointers;
+
+            for (auto f : flat.items()) {
+                json_pointers.push_back(clean_string(f.key()));
+            }
+
+            std::set<std::string> json_pointers_set(json_pointers.begin(), json_pointers.end());
+
+            for (auto s : json_pointers_set) {
+                used_inputs[s] = false;
+            }
+
+            get_prefixes(json_pointers_set);
+        }
+
+        std::string clean_string(std::string str) {
+            bool after_slash = false;
+
+            for (size_t i = 0; i < str.size(); i++) {
+                if (str.at(i) == '/' && i+1 < str.size()) {
+                    if (isdigit(str.at(i+1))) {
+                        after_slash = true;
+                        str.erase(i, 1);
+                        i -= 1;
+                    }
+                }
+                else if (isdigit(str.at(i)) && after_slash) {
+                    str.erase(i, 1);
+                    i -= 1;
+                }
+                else {
+                    after_slash = false;
+                }
+            }
+
+            str.erase(0, 1);
+
[…]
