### 2025-02-11 3024ee6
Update: Updated function calls for the new InputGuard
