### 2025-02-11 6876613
Updated code style of InputGuard, so it matches the rest of the codebase
