+++ src/json_resource.hpp
@@ -16,152 +16,152 @@
 #include <bpstd/string_view.hpp>
 
 struct InputGuard {
-    public:
-        InputGuard(const nlohmann::json &j) {
-            process_json(j);
+  public:
+    InputGuard(const nlohmann::json &j) {
+        process_json(j);
 
-            dict_key_present = false;
-        }
+        this->dict_key_present = false;
+    }
 
-        ~InputGuard() {
-            check_used_inputs();
-        }
+    ~InputGuard() {
+        check_used_inputs();
+    }
 
-        void mark_used_input(const std::string &input_name) {
-            if (prefixes.find(input_name) == prefixes.end()) {
-                std::string prefix = combine_str_vec(curr_prefix);
+    void mark_used_input(const std::string &input_name) {
+        if (this->prefixes.find(input_name) == this->prefixes.end()) {
+            std::string prefix = combine_str_vec(this->curr_prefix);
 
-                if (!prefix.empty()) {
-                    used_inputs[prefix + "/" + input_name] = true;
-                }
-                else {
-                    used_inputs[input_name] = true;
-                }
+            if (!prefix.empty()) {
+                this->used_inputs[prefix + "/" + input_name] = true;
+            }
+            else {
+                this->used_inputs[input_name] = true;
             }
         }
+    }
 
-        void update_dict_key(std::string dict_key) {
-            curr_dict_key = dict_key;
-        }
+    void update_dict_key(std::string dict_key) {
+        this->curr_dict_key = dict_key;
+    }
 
-        void check_read_line(std::string line) {
-            if (line == "mode_name") {
-                if (dict_key_present) {
-                    curr_prefix.pop_back();
-                }
-                curr_prefix.push_back(curr_dict_key);
-                dict_key_present = true;
+    void check_read_line(std::string line) {
+        if (line == "mode_name") {
+            if (this->dict_key_present) {
+                this->curr_prefix.pop_back();
             }
-            else if (line.empty()) {
-                curr_prefix.pop_back();
-                dict_key_present = false;
+            this->curr_prefix.push_back(this->curr_dict_key);
+            this->dict_key_present = true;
+        }
+        else if (line.empty()) {
+            this->curr_prefix.pop_back();
+            this->dict_key_present = false;
+        }
+    }
+
+    void open_spec_file(std::string spec_file_name) {
+        this->curr_prefix.push_back(spec_file_name);
+    }
+
+    void close_spec_file() {
+        this->curr_prefix.pop_back();
+    }
+
+  private:
+    std::map<std::string, bool> used_inputs;
+
+    std::set<std::string> prefixes;
+    std::string curr_dict_key;
+    std::vector<std::string> curr_prefix;
+
+    bool dict_key_present;
+
+    void check_used_inputs() {
+        for (auto item : this->used_inputs) {
+            if (!item.second) {
+                throw std::logic_error(std::string("Failed: \"") + item.first + std::string("\" parameter remains unused."));
             }
         }
+    }
+    
+    void process_json(const nlohmann::json &j) {
+        nlohmann::json flat = j.flatten();
 
-        void open_spec_file(std::string spec_file_name) {
-            curr_prefix.push_back(spec_file_name);
+        // JSON Pointer, as in a string syntax for identifying a specific value in JSON
+        std::vector<std::string> json_pointers;
+
+        for (auto f : flat.items()) {
+            json_pointers.push_back(clean_string(f.key()));
         }
 
-        void close_spec_file() {
-            curr_prefix.pop_back();
+        std::set<std::string> json_pointers_set(json_pointers.begin(), json_pointers.end());
+
+        for (auto s : json_pointers_set) {
+            this->used_inputs[s] = false;
         }
 
-    private:
-        std::map<std::string, bool> used_inputs;
[…]
