### 2025-02-12 #179 @dbaston
Yes, I recommend the package be accepted. I especially appreciate the additions to the first notebook and the creation of a code coverage report.
