### 2025-02-12 #179 @cmarmo
🎉 I am very happy to announce then, that PyPartMC has been approved by pyOpenSci!
Thank you @slayoo and co-authors for submitting PyPartMC and many thanks to @simonom  and @dbaston for reviewing this package! 😸

Please let me know by commenting in the issue or by e-mail if you are interested in an invite to the pyOpenSci Slack.

Also @slayoo, I am going to announce the approval on our [discourse](https://pyopensci.discourse.group/), if you are planning to share your experience with pyOpenSci in a blog post or social posts, please let me know so I can link them to the announcement.

- [x] Add the badge for pyOpenSci peer-review to the README.md of <package-name-here>. The badge should be `[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

---
