### 2025-02-12 7bb3658
add pyOpenSci repo to README.md (#415)
