### 2025-02-12 #179 @slayoo
thank you, @dbaston, @simonom and @cmarmo

✓pyOpenSci badge just added: https://github.com/open-atmos/PyPartMC/pull/415
✓post-review survey filled in
