### 2025-02-16 bf9b4f3
bump Ubuntu CI image (GitHub announced 20.04 will be unsupported by April 1, 2025)
