### 2025-02-17 #179 @slayoo
> I have announced the successful review of PyPartMC on the [pyOpenSci discourse](https://pyopensci.discourse.group/t/welcome-pypartmc-to-the-pyopensci-ecosystem/509).

Thank you, @cmarmo!
